remove raven module and tests
